Version 0.2.2 commit.

+++ iqis.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
-# iqis.sh v0.1
+# iqis
 # Author: Marat Farkhulin (https://iqis.ru) marat.farkhulin@gmail.com
+# Github: https://github.com/farkhulin/iqis_tools
 
 # TODO
 # * DONE Write help page (argument -h)
@@ -10,7 +11,7 @@
 # * DONE Change Restore logic
 
 # * VARIABLES
-VERSION="v0.1"
+VERSION="0.2.2"
 CURRENT=`pwd`
 PROJECT=`basename "$CURRENT"`
 BCKP_SUFFIX="backup"
@@ -26,12 +27,6 @@ DB_NAME="$(php -r 'include("'${DRUPAL_PATH}'sites/default/settings.php"); print 
 DB_USER="$(php -r 'include("'${DRUPAL_PATH}'sites/default/settings.php"); print $databases["default"]["default"]["username"];')"
 DB_PASS="$(php -r 'include("'${DRUPAL_PATH}'sites/default/settings.php"); print $databases["default"]["default"]["password"];')"
 DB_HOST="$(php -r 'include("'${DRUPAL_PATH}'sites/default/settings.php"); print $databases["default"]["default"]["host"];')"
-
-# Testing - REMOVE!
-# echo ${DRUPAL_PATH}
-# echo ${DB_NAME}
-# echo ${PROJECT}
-# echo ${CURRENT}
 
 # * STYLES
 RED='\033[1;31m'
@@ -54,22 +49,25 @@ selfinit() {
         printf "%s\n" "alias iqis='bash ~/.composer/vendor/farkhulin/iqis_tools/iqis.sh'" >> ~/.profile
     fi;
 }
+
 # Function: Display Help.
 usage() {
     echo
-    echo "IQIS.SH - is a script that simplifies the maintenance of your Drupal project."
+    echo -e "${WHITE}IQIS - is a set of scripts that make it easy to maintain your Drupal project."
+    echo -e "For more information visit Github page:"
+    echo -e "https://github.com/farkhulin/iqis_tools"
     echo
-    echo "Syntax: $0 [-a|h|V]"
-    echo "options:"
-    echo "a     action (backup/restore/cleanup)."
-    echo "h     Print this Help."
-    echo "V     Print software version and exit."
+    echo -e "Syntax: iqis [-a|h|V]"
+    echo -e "options:"
+    echo -e "a     Avalibale action (backup/restore/cleanup)."
+    echo -e "h     Print this Help."
+    echo -e "V     Print software version and exit.${NC}"
     echo
 }
 
 # Function: Display Software Version.
 about() {
-    echo -e "IQIS.SH Version: ${VERSION}"
+    echo -e "IQIS Version: ${VERSION}"
     exit 0
 }
 
@@ -81,33 +79,42 @@ exit_abnormal() {
 
 # Function: Backup DB and files full backup.
 backup() {
-    # Run Drush Chache Rebuild if Drush installed.
     cd ${DRUPAL_PATH}
-    if drush cr &> /dev/null
+
+    # Dump DB and files.
+    SETIINGS=${SCRIPT_PATH}sites/default/settings.php
+    if test -f "$SETIINGS"
     then
-        echo -e "${WHITE}Run Drush Chache Rebuild${NC}"
-        drush cr
-        echo -e ""
+        echo -e "A full backup of the site and database will be performed..."
+
+        # Run Drush Chache Rebuild if Drush installed.
+        if drush cr &> /dev/null
+        then
+            echo -e "${WHITE}Run Drush Chache Rebuild${NC}"
+        fi
+
+        echo -e "${WHITE}Create dump ${PROJECT} database${NC}"
+
+        CURRENT_TIME=$( date '+%Y-%m-%d_%H-%M-%S' )
+
+        # Dump DB.
+        mysqldump -u ${DB_USER} -p${DB_PASS} -h ${DB_HOST} ${DB_NAME} | gzip > ${SCRIPT_PATH}_${CURRENT_TIME}_${PROJECT}_db_${BCKP_SUFFIX}.tar.gz
+
+        echo -e "${BG_WHITE}${BLACK} Done ${NC}"
+
+        # Create file archive
+        echo -e "${WHITE}Create ${PROJECT} files archive${NC}"
+
+        touch ${SCRIPT_PATH}_${CURRENT_TIME}_${PROJECT}_files_${BCKP_SUFFIX}.tar.gz
+        tar --exclude=./*.gz --exclude=./*.zip -czf ${SCRIPT_PATH}_${CURRENT_TIME}_${PROJECT}_files_${BCKP_SUFFIX}.tar.gz -C ${DRUPAL_PATH} .
+
+        echo -e "${BG_WHITE}${BLACK} Done ${NC}"
+
+        echo -e "\n${BG_GREEN}${WHITE} All tasks completed ${NC}\n"
+    else
+        echo -e "${WHITE}`pwd` - ${RED}this directory does not contain drupal."
+        echo -e "No database and file archives were created.${NC}"
     fi
-
-    # Dump DB.
-    echo -e "${WHITE}Create dump ${PROJECT} database${NC}"
-
-    CURRENT_TIME=$( date '+%Y-%m-%d_%H-%M-%S' )
-
-    mysqldump -u ${DB_USER} -p${DB_PASS} -h ${DB_HOST} ${DB_NAME} | gzip > ${SCRIPT_PATH}_${CURRENT_TIME}_${PROJECT}_db_${BCKP_SUFFIX}.tar.gz
-
-    echo -e "${BG_WHITE}${BLACK} Done ${NC}"
-
-    # Create file archive
-    echo -e "${WHITE}Create ${PROJECT} files archive${NC}"
-
-    touch ${SCRIPT_PATH}_${CURRENT_TIME}_${PROJECT}_files_${BCKP_SUFFIX}.tar.gz
-    tar --exclude=./*.gz --exclude=./*.zip -czf ${SCRIPT_PATH}_${CURRENT_TIME}_${PROJECT}_files_${BCKP_SUFFIX}.tar.gz -C ${DRUPAL_PATH} .
-
-    echo -e "${BG_WHITE}${BLACK} Done ${NC}"
-
-    echo -e "\n${BG_GREEN}${WHITE} All tasks completed ${NC}\n"
 }
 
 # Function: Restore DB and full files.
@@ -175,25 +182,45 @@ restore() {
 
 # Function: Reset DB and files only iqis.conf and composer.json.
 reset() {
-    echo -e "\n${WHITE}Delete database and all files/directory except iqis.comf and composer.json.${NC}"
+    SETIINGS=${SCRIPT_PATH}sites/default/settings.php
+    if test -f "$SETIINGS"
+    then
+        echo -e "\n${WHITE}Delete database and all files/directory except iqis.conf, composer.json, *.gz and *.zip.${NC}"
 
-    # Clean Database
-    mysql -u ${DB_USER} -p${DB_PASS} -h ${DB_HOST} ${DB_NAME} -e "DROP DATABASE ${DB_NAME}"
-    mysql -u ${DB_USER} -p${DB_PASS} -h ${DB_HOST} -e "CREATE DATABASE ${DB_NAME}"
+        read -p "Are you sure you want to delete the database and all files except the exceptions in this directory: `pwd`? (yes/no) " answer
+        if [ ${answer} = "no" ] ; then
+            echo -e "\nReset aborted.\n"
+        else
+            # Clean Database
+            mysql -u ${DB_USER} -p${DB_PASS} -h ${DB_HOST} ${DB_NAME} -e "DROP DATABASE ${DB_NAME}"
+            mysql -u ${DB_USER} -p${DB_PASS} -h ${DB_HOST} -e "CREATE DATABASE ${DB_NAME}"
 
-    # Clean files/directory
-    chmod -R 777 sites
-    find ./ -type f -not -name 'composer.json' -not -name '*.conf' -delete
-    find ./ -type d -exec rm -rf {} + &> /dev/null
+            # Clean files/directory
+            chmod -R 777 sites
+            find ./ -type f -not -name 'composer.json' -not -name '*.conf' -not -name '*.gz' -not -name '*.zip' -delete
+            find ./ -type d -exec rm -rf {} + &> /dev/null
 
-    echo -e "\n${BG_GREEN}${WHITE} Reset complite! ${NC}\n"
+            echo -e "\n${BG_GREEN}${WHITE} Reset complite! ${NC}\n"
+        fi
+    else
+        echo -e "${WHITE}`pwd` - ${RED}this directory does not contain drupal.${NC}"
+        read -p "Delete database and all files/directory except iqis.conf, composer.json, *.gz and *.zip. in `pwd` anyway? (yes/no) " answer
+        if [ ${answer} = "yes" ] ; then
+            # Clean files/directory
+            chmod -R 777 sites &> /dev/null
+            find ./ -type f -not -name 'composer.json' -not -name '*.conf' -not -name '*.gz' -not -name '*.zip' -delete
+            find ./ -type d -exec rm -rf {} + &> /dev/null
+
+            echo -e "\n${BG_GREEN}${WHITE} Reset files and directory complite! ${NC}\n"
+        fi
+    fi
 }
 
 # Function: Initialisation, enable modules.
 init() {
     echo -e "\n${WHITE}Initialisation, enable modules${NC}"
 
-    read -p 'Assembly type (basic|devel)? : ' assemblytype
+    read -p "Assembly type (basic|devel)? : " assemblytype
 
     if [ ${assemblytype} = "basic" ] ; then
         drush en admin_toolbar, admin_toolbar_tools, module_filter, pathauto, ctools, token, webform, webform_ui, metatag, metatag_views, xmlsitemap, responsive_image, colorbox, imce, mailsystem, swiftmailer, scss_compiler
@@ -270,7 +297,6 @@ do
     case "${options}" in
         a)
             if [ ${OPTARG} = "backup" ] ; then
-                echo -e "a full backup of the site and database will be performed..."
                 backup
             elif [ ${OPTARG} = "restore" ] ; then
                 restore
